+++ data/helpers/index.js
@@ -1,6 +1,7 @@
 export {
   detectProjectType,
   detectProblems,
+  detectUxComplaints,
   detectHumanTasks,
   detectConstraints,
 } from './search.js'